Refactored codebase
Renamed:
- Model -> Simulation

Extracted:
- FunctionPlots
- InsulinPumpModel
- BodyMetabolismModel
- plotly helpers
- form helpers (debouncers)
- compose helper

New:
- Helpers for parsing events (event_parsing)
