+++ REFACTOR.md
@@ -1,0 +1,19 @@
+## Sprint 1
+
+```
+components/
+    App
+    Experiment
+        ExperimentSettings
+        ExperimentGraph
+    FunctionPlots
+pages/
+    _app.js
+model/
+    InsulinPumpModel
+    SimulationModel
+    utils.js
+helpers/
+    plotly
+    form
+```
